Reorganize CalendarIntegration services and fix namespace structure
- Move CalendarExportService and CalendarWorkoutCatalogService from subfolder to Services root
- Move interfaces to existing Services/Interfaces folder
- Update all namespace references from WinUI.Services.CalendarIntegration to WinUI.Services
- Update test files to use new namespace structure
- Remove old CalendarIntegration subfolder
- Build successful with 0 errors
