+++ WinUI/Services/CalendarExportService.cs
@@ -1,8 +1,8 @@
 using System.Text;
 using ClassLibrary.Models;
-using WinUI.Services.CalendarIntegration.Interfaces;
+using WinUI.Services.Interfaces;
 
-namespace WinUI.Services.CalendarIntegration;
+namespace WinUI.Services;
 
 public sealed class CalendarExportService : ICalendarExportService
 {
